rename libcloud.httplib_ssl to libcloud.http since it has nothing to do with httplib anymore
