+++ libcloud/http.py
